+++ system/core/Output.php
@@ -64,7 +64,7 @@ class CI_Output {
 	 *
 	 * @var array
 	 */
-	public $mime_types =	array();
+	public $mimes =		array();
 
 	/**
 	 * Determines wether profiler is enabled
@@ -104,14 +104,7 @@ class CI_Output {
 		$this->_zlib_oc = (bool) @ini_get('zlib.output_compression');
 
 		// Get mime types for later
-		if (defined('ENVIRONMENT') && file_exists(APPPATH.'config/'.ENVIRONMENT.'/mimes.php'))
-		{
-			$this->mime_types = include APPPATH.'config/'.ENVIRONMENT.'/mimes.php';
-		}
-		else
-		{
-			$this->mime_types = include APPPATH.'config/mimes.php';
-		}
+		$this->mimes =& get_mimes();
 
 		log_message('debug', 'Output Class Initialized');
 	}
@@ -214,9 +207,9 @@ class CI_Output {
 			$extension = ltrim($mime_type, '.');
 
 			// Is this extension supported?
-			if (isset($this->mime_types[$extension]))
+			if (isset($this->mimes[$extension]))
 			{
-				$mime_type =& $this->mime_types[$extension];
+				$mime_type =& $this->mimes[$extension];
 
 				if (is_array($mime_type))
 				{